requirements.txt
# Core web scraping and automation
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
urllib3>=2.0.0

# HTML parsing and processing
lxml>=4.9.0
html5lib>=1.1

# Database and data handling
sqlalchemy>=2.0.0
psutil>=5.9.0

# Cloud storage (AWS S3)
boto3>=1.34.0
botocore>=1.34.0

# MinIO storage
minio>=7.2.0

# Web framework (FastAPI)
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# Task queue
celery>=5.3.4
redis>=5.0.1

# Database
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv>=1.0.0

# HTTP and networking
certifi>=2023.7.0
charset-normalizer>=3.3.0
idna>=3.4

# Data processing and utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: For enhanced functionality
# Pillow>=10.0.0  # For image processing (uncomment if needed)
# pandas>=2.0.0   # For data analysis (uncomment if needed)
# numpy>=1.24.0   # For numerical operations (uncomment if needed)

# Development and testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
